requirements.txt
numpy>=1.24
pandas>=2.0
yfinance>=0.2.40
scikit-learn>=1.3
matplotlib>=3.8
pyyaml>=6.0
pytest>=7.4
